Further progress on UART bluetooth comms
We actually manage to send a few things back and forth now,
including the (for now fixed) MAC address of the controller.

It's not fast enough yet though. The overrun flag (PH3) is set after
a while, and we can also see that we are only receiving a mangled part
of the MAC address from bluetooth.